timechange-package {timechange}	R Documentation

Package timechange

Description

Utilities for efficient updating of date-times components while accounting for time-zones and day-light saving times. When it makes sense functions provide a refined control of what happens in ambiguous situations through roll_month and roll_dst arguments.

Author(s)

Vitalie Spinu (spinuvit@gmail.com)

See Also

Useful links:

https://github.com/vspinu/timechange/

Report bugs at https://github.com/vspinu/timechange/issues

[Package timechange version 0.1.1 Index]